+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/caoyingjunz/gopixiu
 
-go 1.19
+go 1.16
 
 require (
 	github.com/bndr/gojenkins v1.1.0
